Updated Checkout1.jsp
Changed the color of the input fields to black In this CSS, set the width of the .form-group class to 95% to allow for more space for input fields. Additionally, I've set the width of input fields, select elements, and textareas to 100%, ensuring they span the entire width of their container.